Tests: Don't preserve the global state for Ajax tests when using an external object cache.
Most of the Ajax tests are running in a separate PHP process, and thus PHPUnit attempts to preserve the global state from the parent process by serializing all globals. But this doesn't work for external object caches so we have to disable this "feature".

See #31491.
Built from https://develop.svn.wordpress.org/trunk@35209


git-svn-id: http://core.svn.wordpress.org/trunk@35175 1a063a9b-81f0-0310-95a4-ce76da25c4cd