HTTP API: Ensure the `http_api_debug` hook is fired for all responses.
This means that logging and debugging functionality can access the error code and error message for erroneous responses under all conditions. The hook is not fired when a request is short-circuited with the `pre_http_request` filter, because this filter itself can be used in that situation.

Fixes #25747

Built from https://develop.svn.wordpress.org/trunk@45504


git-svn-id: http://core.svn.wordpress.org/trunk@45315 1a063a9b-81f0-0310-95a4-ce76da25c4cd