Build/Test Tools: Run automated testing workflows for `trunk` branch.
This ensures a seamless transition when the primary branch is changed from `master` to `trunk` in the GitHub mirror. Once the transition is completed, the `master` branch can be removed.

See https://make.wordpress.org/systems/2021/03/02/update-git-wordpress-org-to-rename-master-to-trunk/ and https://make.wordpress.org/core/2020/06/18/proposal-update-all-git-repositories-to-use-main-instead-of-master/.

See #52625.
Built from https://develop.svn.wordpress.org/trunk@50473


git-svn-id: http://core.svn.wordpress.org/trunk@50084 1a063a9b-81f0-0310-95a4-ce76da25c4cd